+++ app/src/main/kotlin/me/xizzhu/android/joshua/search/result/SearchResultPresenter.kt
@@ -19,12 +19,13 @@ package me.xizzhu.android.joshua.search.result
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.consumeEach
 import kotlinx.coroutines.launch
-import me.xizzhu.android.joshua.core.Verse
 import me.xizzhu.android.joshua.core.VerseIndex
 import me.xizzhu.android.joshua.core.logger.Log
 import me.xizzhu.android.joshua.search.SearchInteractor
 import me.xizzhu.android.joshua.ui.LoadingSpinnerState
+import me.xizzhu.android.joshua.ui.recyclerview.BaseItem
 import me.xizzhu.android.joshua.ui.recyclerview.SearchItem
+import me.xizzhu.android.joshua.ui.recyclerview.TitleItem
 import me.xizzhu.android.joshua.utils.BaseSettingsPresenter
 import java.util.ArrayList
 
@@ -35,8 +36,13 @@ class SearchResultPresenter(private val searchInteractor: SearchInteractor)
 
         launch(Dispatchers.Main) {
             searchInteractor.observeSearchResult().consumeEach { (query, verses) ->
-                val items = ArrayList<SearchItem>()
+                val items = ArrayList<BaseItem>()
+                var currentBookIndex = -1
                 verses.forEach { verse ->
+                    if (currentBookIndex != verse.verseIndex.bookIndex) {
+                        items.add(TitleItem(verse.text.bookName))
+                        currentBookIndex = verse.verseIndex.bookIndex
+                    }
                     items.add(SearchItem(verse.verseIndex, verse.text.bookName,
                             verse.text.text, query, this@SearchResultPresenter::selectVerse))
                 }