+++ CHANGELOG.md
@@ -5,6 +5,10 @@ CHANGELOG
 - Supported sorting bookmarks and notes by book
 - Divided search results by books
 - Supported adaptive icons
+- Updated dependencies:
+  - AndroidX JUnit to 1.1.1
+  - AndroidX Test Rules to 1.2.0
+  - Crashlytics to 2.10.1
 
 #### v0.4.1 (2019-05-23)
 - Added missing on click listener for simple reading mode